+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Sections settings icon to show the section name when hovered: When hovering an the gear icon of a section, have its name showing #143 
 
+- Hide sections settings icons when side bar is opened: When adding or editing a section, the site side must be as if I am in view mode #142
 
 ### [1.0.17] - 2025-02-03