+++ CHANGELOG.md
@@ -11,12 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for custom sections 404 page that its path can be set through Sections BO: Ability to create a page on section for a project and use it when we have a 404 page not found #145
 
-
 ### Updated
 
 - Sections settings icon to show the section name when hovered: When hovering an the gear icon of a section, have its name showing #143 
 
 - Hide sections settings icons when side bar is opened: When adding or editing a section, the site side must be as if I am in view mode #142
+
+- Upgrade vue-components library version to v1.0.12
 
 ### Fixed